Fix encode and decode on 32-bit targets
The entropy bit accumulator was `usize`; the format needs 64 bits. Make it
`u64` on every target, clamp the parameter-driven tables that a 32-bit
address space cannot hold, and add CI legs that would have caught it:
wasm32-wasip1 for the suite, i686 under Miri for the bounds arithmetic.

+++ src/asm_inspect.rs
@@ -44,7 +44,7 @@ use crate::entropy::fse::{DState, SequenceDecodeEntry};
 /// +2-3 insns (extra shift + sub + and for the zero-safe mask).
 #[inline(never)]
 #[unsafe(no_mangle)]
-pub extern "C" fn asm_read_bits_fast_zero_safe(stream: &mut BitDStream, nb_bits: u32) -> usize {
+pub extern "C" fn asm_read_bits_fast_zero_safe(stream: &mut BitDStream, nb_bits: u32) -> u64 {
     stream.read_bits_fast_zero_safe(nb_bits)
 }
 
@@ -52,14 +52,14 @@ pub extern "C" fn asm_read_bits_fast_zero_safe(stream: &mut BitDStream, nb_bits:
 /// Should produce identical instruction count to C.
 #[inline(never)]
 #[unsafe(no_mangle)]
-pub extern "C" fn asm_read_bits_fast(stream: &mut BitDStream, nb_bits: u32) -> usize {
+pub extern "C" fn asm_read_bits_fast(stream: &mut BitDStream, nb_bits: u32) -> u64 {
     stream.read_bits_fast(nb_bits)
 }
 
 /// Rust's look_bits_fast (no skip, matches C BIT_lookBitsFast).
 #[inline(never)]
 #[unsafe(no_mangle)]
-pub extern "C" fn asm_look_bits_fast(stream: &BitDStream, nb_bits: u32) -> usize {
+pub extern "C" fn asm_look_bits_fast(stream: &BitDStream, nb_bits: u32) -> u64 {
     stream.look_bits_fast(nb_bits)
 }
